+++ cyclope/default_urls.py
@@ -93,13 +93,6 @@ if django_settings.DEBUG:
             {'document_root': django_settings.MEDIA_ROOT, 'show_indexes': True})
     )
 
-    import os, feincms
-    feincms_root = os.path.join(feincms.__path__[0], 'media/feincms/')
-    urlpatterns+= patterns('',
-        url(r'^feincms_media/(?P<path>.*)$', 'django.views.static.serve',
-            {'document_root': feincms_root, 'show_indexes': True})
-    )
-
 if 'live' in django_settings.INSTALLED_APPS:
     urlpatterns += patterns('',
                             (r'^live/', include('live.urls')),